requirements.txt
opencv-python==4.6.0.66
numpy==2.4.2
Pillow==9.2.0
requests==2.32.5